EMF: Implicit channel values start at 16, not 17

+++ src/EventManifestFramework/EventManifestParser.cs
@@ -435,7 +435,7 @@ namespace EventManifestFramework
 
                 channel.Index = provider.Channels.Count;
                 if (channel.Value == null)
-                    channel.Value = (byte)provider.CreateChannelValue();
+                    channel.Value = provider.CreateChannelValue();
 
                 provider.Channels.TryAdd(channel, diags);
             }
